pyproject.toml
[project]
name = "infiniquantumsim"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "duckdb==1.1.3",
    "jupyter-client==8.6.3",
    "jupyter-core==5.7.2",
    "numpy==2.2.5",
    "opt-einsum==3.3.0",
    "psycopg2-binary==2.9.10",
    "pytest>=9.0.2",
]

[tool.setuptools.packages.find]
include = ["InfiniQuantumSim*"]
exclude = ["results*"]
